refactor: group 엔티티 이름 team으로 변경
